Export `undefined` from Lesson2_Part2 prelude.
It was not part of Lesson2_Part2's prelude, despite being exported in
the preludes from prior exercise modules, as it wasn't obviously
needed to solve the exercises in Lesson2_Part2, but we've decided to
change our pedagogical approach on prelude design.

Going forward, unless there are exceptional circumstances, successive
preludes should take care to be a superset of their predecessors. Our
rationale for this is as follows:

- Principle of least surprise (see
  https://en.wikipedia.org/wiki/Principle_of_least_astonishment)

- Students may find ways to use previously-encountered functions to
  solve exercises in ways we, the course designers, did not
  anticipate. Therefore, we should not presume to know that these
  previously-encountered functions won't be useful to solve a
  particular set of exercises.

- Students' first inclination might be to lean on these
  previously-encountered functions when struggling to understand a new
  concept, only to find that they don't help. We should enable this
  form of trial and error by giving students the "tools to fail," so
  to speak, where "failure" is defined by trying something that
  previously worked and now, under different conditions or
  assumptions, does not.