Lesson2_Part3.hs: add function use exercises.
Some of these use the evaluation game. Once we have a stable version
and a URL, we can put that URL in the exercise docs so that the
student knows how to load that exercise up in the game.

The last exercise is a general recursion test! I believe it should be
relatively easy for the student to reason about, so long as they have
a good understanding of evaluation by the time they get to this
exercise.