Remove Exercise 4.10.
I think the intent of this exercise is good. However, I'm concerned
that the cognitive load here is a bit too much to handle at this stage
of the course. If we continue developing the course, I think it'd be a
good candidate for a later lesson where we want to revisit recursion
in a spaced repetition-type way.