Remove Exercise 4.9.
I don't believe it's possible to write this function without (==),
which we haven't covered in any of the lessons yet as we haven't
needed it.

We could try to introduce it here in the exercise, but I haven't given
that much thought, and as I'm under a bit of time pressure, I'd need
to do a "rush job," so to speak.

Also, even if we did introduce it here, (==) requires Bool from base,
not the Bool we've defined in our lessons (including in the Lesson4
exercises), so we'd have to explain that, as well.

Anyway, it's a good exercise and, as with the previous commit that
removed Exercise 4.10, we should probably include it in the exercises
in a later lesson, after we've dealt with (==) and Eq (and start using
Bool from base).