Remove Exercise 4.11.
(Normally I would PR this, but I'm under time pressure to get these
exercises ready, and Harry won't be back before the student does the
exercises.)

I believe this requires inequalities (which we haven't introduced
yet). Alternatively we could introduce a `max` function, but that
should probably be done in a lesson, not here.

I think it's probably a bit confusing to talk about the height of the
tree when we're starting from the root at the top of the drawing. I
don't object to drawing the tree that way, of course, but I wonder if
we shouldn't call this function "depth".

I also have small quibble about the drawing: the number marks on the
ruler on the right-hand side look a bit like negative numbers, which
might lead to confusion.